backend/requirements.txt
fastapi
uvicorn
structlog
pydantic-settings
anthropic
openai
google-cloud-bigquery
weaviate-client
pymongo
redis
psycopg2-binary
neo4j
sqlparse
pandas
numpy
openpyxl
xlrd
sqlalchemy
aioredis
motor
aiohttp
google-cloud-storage
psutil
pytesseract
pillow
scikit-learn
matplotlib
seaborn
plotly
pypdf
python-docx
python-multipart
